Fix compilation warnings on MSVC, fix coverall test compilation issue on MSVC

+++ bindgen/src/bindings/cpp/templates/arith_conv.cpp
@@ -18,5 +18,5 @@ void {{ ffi_converter_name }}::write(RustStream &stream, {{ type_name }} val) {
 }
 
 int32_t {{ ffi_converter_name }}::allocation_size({{ type_name }}) {
-    return sizeof({{ type_name }});
+    return static_cast<int32_t>(sizeof({{ type_name }}));
 }
